# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Education Experiment Platform API",
        "version": "2.0.0",
        "x-openapi-revision": "2026-09-20.b6-pending-freeze",
        "x-generated-at": "2026-09-20T17:07:29+08:00",
        "x-contract-status": "pending-freeze; replace after contract freeze"
    },
    "paths": {
...
        "/api/v2/tasks/{public_id}/workspace-sessions": {
            "post": {
                "operationId": "api_v2_tasks_workspace_sessions_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "public_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "workspaces"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkspaceSessionCreateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkspaceSessionCreateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkspaceSessionCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkspaceSessionCreateResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "WorkspaceSessionCreateRequest": {
                "type": "object",
                "properties": {
                    "student_public_id": {
                        "type": "string"
                    },
                    "instance_public_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "instance_public_id",
                    "student_public_id"
                ]
            },
            "WorkspaceSessionCreateResponse": {
                "type": "object",
                "properties": {
                    "session_public_id": {
                        "type": "string"
                    },
                    "task_public_id": {
                        "type": "string"
                    },
                    "student_public_id": {
                        "type": "string"
                    },
                    "instance_public_id": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "replayed": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "instance_public_id",
                    "replayed",
                    "session_public_id",
                    "student_public_id",
                    "task_public_id"
                ]
            },
...
        }
    }
}